EMB-1716: refactor(observability): convert reporter to class
Match peer convention (Logger, ApiErrorHandler, Configuration). Replace the
exported const-object with a class exposing a static reportError method.
Drop the unnecessary `as { componentStack: unknown }` cast in favor of a
hasStringComponentStack type guard. Rename `args` to `loggerArgs` at the
reporter boundary for clearer intent.
